Fix for too long G-code replies (#1073)
Added extra parameter to OutputBuffer::Allocate to specify if the buffer
to be allocated may be taken from the output buffer reserve. This is
always true except if appending to an existing output buffer chain or
when the request may be repeated without downsides.
Also reduced the number of reserved output buffers to 2 on Duet 2 series